pyproject.toml
[project]
name = "nakari"
version = "0.1.0"
description = "A unique AI agent that develops personality through accumulated experiences"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.60.0",
    "neo4j>=5.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "rich>=13.0.0",
]

[project.scripts]
nakari = "nakari.cli:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.targets.wheel]
packages = ["nakari"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
indent-width = 2
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
quote-style = "double"
